move graph transform functions
