+++ src/app/ranking/components/ContestLogsTable.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { ContestLog } from '../interfaces'
 import styled from 'styled-components'
 import { languageNameByCode, mediumDescriptionById } from '../database'
-import { amountToPages } from '../transform'
+import { amountToPages } from '../transform/graph'
 import { Button, ButtonContainer } from '../../ui/components'
 import media from 'styled-media-query'
